# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorCode": {
        "description": "Stable API error codes consumed by the web client.",
        "enum": [
          "VALIDATION_ERROR",
          "EMPTY_FILE",
          "FILE_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "INVALID_ENCODING",
          "MALFORMED_FDX",
          "MALFORMED_PDF",
          "INVALID_NATIVE_DOCUMENT",
          "PROCESS_RESTARTED",
          "UNAUTHENTICATED",
          "AUTHENTICATION_UNAVAILABLE",
          "ACCOUNT_BINDING_CONFLICT",
          "DOCUMENT_NOT_FOUND",
          "TASK_NOT_FOUND",
          "NOT_FOUND",
          "FORBIDDEN",
          "CONFLICT",
          "DOCUMENT_CONTENT_MISSING",
          "CONTRACT_MISMATCH",
          "INTERNAL_ERROR"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "CreateVoiceRequest": {
        "description": "Request payload to create a new voice persona.",
        "properties": {
          "language_code": {
            "default": "ja-JP",
            "description": "Language code",
            "title": "Language Code",
            "type": "string"
          },
          "speaker": {
            "description": "Speaker name (narrator or character)",
            "minLength": 1,
            "title": "Speaker",
            "type": "string"
          },
          "voice_traits": {
            "description": "Natural language voice description",
            "minLength": 1,
            "title": "Voice Traits",
            "type": "string"
          }
        },
        "required": [
          "speaker",
          "voice_traits"
        ],
        "title": "CreateVoiceRequest",
        "type": "object"
      },
      "CreateVoiceResponse": {
        "description": "Response payload containing created voice assignment details.",
        "properties": {
          "speaker": {
            "title": "Speaker",
            "type": "string"
          },
          "voice_id": {
            "title": "Voice Id",
            "type": "string"
          },
          "voice_traits": {
            "title": "Voice Traits",
            "type": "string"
          }
        },
        "required": [
          "speaker",
          "voice_id",
          "voice_traits"
        ],
        "title": "CreateVoiceResponse",
        "type": "object"
      },
...
      "DocumentItem": {
        "description": "Summary item for document list view.",
        "properties": {
          "arc_distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arc Distance"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_version_id": {
            "title": "Current Version Id",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "expected_version": {
            "title": "Expected Version",
            "type": "integer"
          },
          "gcs_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gcs Uri"
          },
          "is_saved": {
            "title": "Is Saved",
            "type": "boolean"
          },
          "owner_email": {
            "title": "Owner Email",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "string"
          },
          "shared_count": {
            "title": "Shared Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "owner_user_id",
          "title",
          "current_version_id",
          "version_id",
          "expected_version",
          "is_saved",
          "created_at",
          "updated_at",
          "owner_email",
          "shared_count"
        ],
        "title": "DocumentItem",
        "type": "object"
      },
      "DocumentListResponse": {
        "description": "Paginated document list response.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
...
      "VoicePreviewRequest": {
        "description": "Request payload to synthesize a short sample line with an existing voice.",
        "properties": {
          "language_code": {
            "default": "ja-JP",
            "description": "Language code",
            "title": "Language Code",
            "type": "string"
          },
          "style": {
            "default": "",
            "description": "Optional performance direction",
            "title": "Style",
            "type": "string"
          },
          "text": {
            "description": "Short sample line to synthesize",
            "maxLength": 500,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "voice_id": {
            "description": "Created voice id (voice_...)",
            "minLength": 1,
            "title": "Voice Id",
            "type": "string"
          }
        },
        "required": [
          "voice_id",
          "text"
        ],
        "title": "VoicePreviewRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Intelligent screenplay analysis and narrative arc workbench API",
    "title": "NARRAVANT API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/documents": {
      "get": {
        "description": "List screenplays with optional filtering and narrative arc similarity matching.",
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "description": "Search term for title",
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for title",
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Repeated field:direction server-side sort terms",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeated field:direction server-side sort terms",
              "title": "Sort"
            }
          },
          {
            "description": "Document ID to find similar narrative arcs",
            "in": "query",
            "name": "similar_to_arc_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Document ID to find similar narrative arcs",
              "title": "Similar To Arc Id"
            }
          },
          {
            "description": "Configured Valence pattern ID to find similar narrative arcs; mutually exclusive with similar_to_arc_id",
            "in": "query",
            "name": "arc_pattern_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Configured Valence pattern ID to find similar narrative arcs; mutually exclusive with similar_to_arc_id",
              "title": "Arc Pattern Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/voices": {
      "post": {
        "description": "Create a custom voice persona using Gemini Voice Design.\n\nNOTE (MF-2): This endpoint calls Voice Design and returns the created voice_id.\nIt does NOT persist to the document or advance version_id. Persistence of\nvoice_assignments is handled via regular PUT /api/v1/documents/{doc_id}.",
        "operationId": "create_document_voice_api_v1_documents__document_id__voices_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Document Voice",
        "tags": [
          "voices"
        ]
      }
    },
    "/api/v1/documents/{document_id}/voices/preview": {
      "post": {
        "description": "Synthesize a short sample line with an existing voice for the check button.\n\nReturns raw PCM audio (audio/l16; rate=24000; channels=1). The sample line is\nnot persisted; it is used only to preview a created voice.",
        "operationId": "preview_document_voice_api_v1_documents__document_id__voices_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoicePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Preview Document Voice",
        "tags": [
          "voices"
        ]
      }
    },
...
    "/api/v1/voices": {
      "post": {
        "description": "Create a voice for an unsaved import draft without creating a document.",
        "operationId": "create_draft_voice_api_v1_voices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Draft Voice",
        "tags": [
          "voices"
        ]
      }
    },
    "/api/v1/voices/preview": {
      "post": {
        "description": "Preview a voice assigned only in an unsaved import draft.",
        "operationId": "preview_draft_voice_api_v1_voices_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoicePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Preview Draft Voice",
        "tags": [
          "voices"
        ]
      }
    },
...
  }
}